requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.6
django-cleanup==8.1.0
django-tinymce==4.1.0
idna==3.7
oauthlib==3.2.2
pillow==10.4.0
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
